+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+v5.7.55
+----------
+ * Use v13 flow as example on definitions endpoint docs
+ * Add URNs field to FlowStart and pass to mailroom so that it creates contacts
+
+v5.7.54
+----------
+ * Update editor to get support for expressions in add to group actions
+ * Remove unused localized text on Msg and Broadcast
+
+v5.7.52
+----------
+ * Migrations and models for new imports
+
 v5.7.51
 ----------
  * Add plan_start, calculate active contacts in plan period, add to OrgActivity